Initial commit
This ports ports and refactors the blog engine used for fred.ca
so that it can be used for more general projects.

This refactoring also adds basic support for markdown syntax to
be used within the ErlyDTL templates by wrapping them in
{% markdown %} {% endmarkdown %} tags (which can not be nested).

Moreover, I've included a build system to create a release so
custom binaries can be generated for whoever wants them, or just
dependencies to be used with an existing Erlang install.

The demo blog (in the demo-blog) directory shows a basic blog
config and generation to be had to get things going.

To compile, have agner and an agnerized rebar installed (agner
install rebar), and then do:
  rebar get-deps
  rebar compile
To build dependencies, go into rel/ and run
  ./build.sh
The resulting erlang release is called in demo-blog/run.sh and will
build the demo blog app with the release.