Abstracting away titles and dates
This fix makes it so titles and dates for blog entries can be
injected straight in the base template and overwritten on demand.
The values are taken from the blog's index to avoid duplication of
information.

This introduces reserved variables under the name 'meta' that will
be injected by blogerl into the namespace we have.